socketapp/requirements.txt
motor==2.5.1
pymongo==3.12.0
python-dotenv==0.19.0
websockets==10.0
